Addresses the "show max on bar" preference not loading the first time you log in
Currencies can now give the simple currency plugin a hint to configure the currency as if it has a max value, since when you first load, the value you get from the WoW API can say it doesn't have a max.

Also fixed the "can get" amount on the tooltip, since seasonal currencies have a moving maximum.

Also added the "can get" value to the bar when showing the max value, currently this value is in square brackets "[ ]"