Quick fix after the Titan Panel 8.0.6 changes
The Titan Panel 8.0.6 removed the function TITAN_PANEL_MENU_CATEGORIES used to this addon to create a menu categorization, which means that I can only use the pre-builded menus made available by Titan Panel. For now, all the currencies plugins will be available at the menu Information, and the item currencies wil be in the menu Interface.